For compressed meshes missing m_VertexCount, detect UV / normal maps
Sorry if this ends up not being the best approach to this, I'm still very much a beginner dealing with mesh data.

I came across some unity assets with m_CompressedMesh data, including UV maps and normal maps, but for whatever reason m_VertexCount is reported as zero. Previously, this would cause UV/normal map detection to be skipped.

I couldn't find any other info in the asset file that could definitively point us to 2 dimensions or 3 in this case, but it seemed like we could make a pretty high quality guess based on the bounding box. So I added this logic as a fallback when this scenario happens, then the UV/normal checks are able to proceed, and I've been having good results with this on my local copy.